# perf(tree): N*log(N) composition of sequence-shaped changesets