--- a/packages/dds/tree/src/test/util/utils.spec.ts
+++ b/packages/dds/tree/src/test/util/utils.spec.ts
@@ -6,6 +6,7 @@
 import { strict as assert } from "node:assert";
 
 import {
+	balancedReduce,
 	capitalize,
 	copyProperty,
 	defineLazyCachedProperty,
@@ -94,6 +95,66 @@ describe("Utils", () => {
 			const destination = {};
 			copyProperty(source, "a", destination);
 			assert.equal(Reflect.has(destination, "a"), false);
+		});
+	});
+
+	describe("balancedReduce", () => {
+		let delegateCallCount = 0;
+		const concatDelegate = (a: string, b: string) => {
+			delegateCallCount += 1;
+			return a + b;
+		};
+
+		let factoryCallCount = 0;
+		const factory = () => {
+			factoryCallCount += 1;
+			return "factory";
+		};
+
+		beforeEach(() => {
+			factoryCallCount = 0;
+			delegateCallCount = 0;
+		});
+
+		it("uses empty case factory on empty input", () => {
+			const actual = balancedReduce([], concatDelegate, factory);
+			assert.equal(actual, "factory");
+			assert.equal(factoryCallCount, 1);
+			assert.equal(delegateCallCount, 0);
+		});
+
+		it("returns lone element on size 1 input", () => {
+			const actual = balancedReduce(["lone"], concatDelegate, factory);
+			assert.equal(actual, "lone");
+			assert.equal(factoryCallCount, 0);
+			assert.equal(delegateCallCount, 0);
+		});
+
+		it("calls delegate once on size 2 input", () => {
+			const actual = balancedReduce(["A", "B"], concatDelegate, factory);
+			assert.equal(actual, "AB");
+			assert.equal(factoryCallCount, 0);
+			assert.equal(delegateCallCount, 1);
+		});
+
+		it("calls delegate twice on size 3 input", () => {
+			const actual = balancedReduce(["A", "B", "C"], concatDelegate, factory);
+			assert.equal(actual, "ABC");
+			assert.equal(factoryCallCount, 0);
+			assert.equal(delegateCallCount, 2);
+		});
+
+		it("calls delegate with balanced inputs", () => {
+			const delegate = (a: string, b: string) => {
+				delegateCallCount += 1;
+				// Checks that the inputs are balanced.
+				assert(Math.abs(a.length - b.length) <= 1);
+				return a + b;
+			};
+			const actual = balancedReduce(["A", "B", "C", "E", "F", "G", "H"], delegate, factory);
+			assert.equal(actual, "ABCEFGH");
+			assert.equal(factoryCallCount, 0);
+			assert.equal(delegateCallCount, 6);
 		});
 	});
 });