## Description

Uses a more efficient algorithm for composing sequence-shaped changesets (both sequence and generic field kinds).
It benefits from the same principle that makes merge sort faster than insertion sort, making it O(N*log(N)) instead of O(N²) for N changesets each containing 1 change atom.

Times from the included benchmark:
| Scenario                                              | Before PR (ns) | After PR  (ns) | Reduction (%) |
| ----------------------------------------------------- | -------------- | -------------- | ------------- |
| Compose 10 sequence edits into a single transaction   | 2,141,503      | 2,053,101      | 4%            |
| Compose 100 sequence edits into a single transaction  | 28,166,687     | 16,318,033     | 42%           |
| Compose 1000 sequence edits into a single transaction | 1,490,000,000  | 169,982,458    | 89%           |

## Breaking Changes

None